#10 refactor : Comment
bulider 패턴으로 entity에 정적 메소드 생성
recode클래스에 있던 메서드 삭제
responseDto추가
