+++ src/main/java/com/sparta/jamrello/domain/comment/dto/CommentResponseDto.java
@@ -1,0 +1,7 @@
+package com.sparta.jamrello.domain.comment.dto;
+
+public record CommentResponseDto(
+    String content
+) {
+
+}
